+++ src/main/java/gift/member/controller/AuthController.java
@@ -1,0 +1,30 @@
+package gift.member.controller;
+
+import gift.member.dto.LoginRequestDto;
+import gift.member.dto.LoginResponseDto;
+import gift.member.service.AuthService;
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/members/")
+public class AuthController {
+
+    private final AuthService authService;
+
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<LoginResponseDto> login(
+        @Valid @RequestBody LoginRequestDto loginRequestDto) {
+
+        return new ResponseEntity<>(authService.login(loginRequestDto), HttpStatus.OK);
+    }
+}
