+++ src/main/java/gift/member/service/AuthService.java
@@ -1,0 +1,10 @@
+package gift.member.service;
+
+import gift.member.dto.LoginRequestDto;
+import gift.member.dto.LoginResponseDto;
+
+public interface AuthService {
+
+    LoginResponseDto login(LoginRequestDto loginRequestDto);
+
+}
