+++ src/main/java/gift/member/service/MemberServiceImpl.java
@@ -1,7 +1,5 @@
 package gift.member.service;
 
-import gift.member.dto.LoginRequestDto;
-import gift.member.dto.LoginResponseDto;
 import gift.member.dto.RegisterRequestDto;
 import gift.member.dto.RegisterResponseDto;
 import gift.member.entity.Member;
@@ -21,16 +19,6 @@ public class MemberServiceImpl implements MemberService {
     public RegisterResponseDto registerMember(RegisterRequestDto registerRequestDto) {
         Member member = new Member(registerRequestDto.email(), registerRequestDto.password());
         memberRepository.saveMember(member);
-        return null;
-    }
-
-    @Override
-    public LoginResponseDto login(LoginRequestDto loginRequestDto) {
-        // DB 조회 -> 성공 시 Token 생성, 실패 시 로그인 실패
-        memberRepository.findMemberByEmail(loginRequestDto.email());
-
-        // TODO: 성공 시 Token 생성
-
         return null;
     }
 }
